+++ Sources/CircularProgress/Vendored/CustomButton.swift
@@ -9,9 +9,9 @@ open class CustomButton: NSButton {
 	public static func circularButton(title: String, radius: Double, center: CGPoint) -> CustomButton {
 		with(CustomButton()) {
 			$0.title = title
-			$0.frame = CGRect(x: Double(center.x) - radius, y: Double(center.y) - radius, width: radius * 2, height: radius * 2)
+			$0.frame = CGRect(x: center.x - radius, y: center.y - radius, width: radius * 2, height: radius * 2)
 			$0.cornerRadius = radius
-			$0.font = .systemFont(ofSize: CGFloat(radius * 2 / 3))
+			$0.font = .systemFont(ofSize: radius * 2 / 3)
 		}
 	}
 
@@ -209,9 +209,9 @@ open class CustomButton: NSButton {
 
 	override open func updateLayer() {
 		let isOn = state == .on
-		layer?.cornerRadius = CGFloat(cornerRadius)
-		layer?.borderWidth = CGFloat(borderWidth)
-		layer?.shadowRadius = CGFloat(isOn && activeShadowRadius != -1 ? activeShadowRadius : shadowRadius)
+		layer?.cornerRadius = cornerRadius
+		layer?.borderWidth = borderWidth
+		layer?.shadowRadius = isOn && activeShadowRadius != -1 ? activeShadowRadius : shadowRadius
 		layer?.shadowOpacity = Float(isOn && activeShadowOpacity != -1 ? activeShadowOpacity : shadowOpacity)
 		animateColor()
 	}
